[refactor/#87]: StatisticsRepository 파일 이동
 - 커스텀 레파지토리랑 같이 패키지에 넣기 위해
