+++ src/main/java/com/woohaengshi/backend/service/auth/AuthServiceImpl.java
@@ -19,7 +19,7 @@ import com.woohaengshi.backend.dto.result.SignInResult;
 import com.woohaengshi.backend.exception.WoohaengshiException;
 import com.woohaengshi.backend.repository.MemberRepository;
 import com.woohaengshi.backend.repository.RefreshTokenRepository;
-import com.woohaengshi.backend.repository.StatisticsRepository;
+import com.woohaengshi.backend.repository.statistics.StatisticsRepository;
 import com.woohaengshi.backend.repository.SubjectRepository;
 
 import jakarta.transaction.Transactional;
