+++ src/main/java/com/woohaengshi/backend/service/statistics/StatisticsScheduler.java
@@ -1,6 +1,6 @@
 package com.woohaengshi.backend.service.statistics;
 
-import com.woohaengshi.backend.repository.StatisticsRepository;
+import com.woohaengshi.backend.repository.statistics.StatisticsRepository;
 
 import lombok.RequiredArgsConstructor;
