+++ src/main/java/com/woohaengshi/backend/service/statistics/StatisticsServiceImpl.java
@@ -9,8 +9,7 @@ import com.woohaengshi.backend.dto.response.statistics.RankDataResponse;
 import com.woohaengshi.backend.dto.response.statistics.ShowRankSnapshotResponse;
 import com.woohaengshi.backend.exception.ErrorCode;
 import com.woohaengshi.backend.exception.WoohaengshiException;
-import com.woohaengshi.backend.repository.StatisticsRepository;
-import com.woohaengshi.backend.repository.StatisticsSpecification;
+import com.woohaengshi.backend.repository.statistics.StatisticsRepository;
 import com.woohaengshi.backend.repository.StudyRecordSpecification;
 import com.woohaengshi.backend.repository.studyrecord.StudyRecordRepository;
