+++ src/main/java/com/woohaengshi/backend/service/studyrecord/StudyRecordServiceImpl.java
@@ -16,7 +16,7 @@ import com.woohaengshi.backend.dto.response.studyrecord.ShowYearlyRecordResponse
 import com.woohaengshi.backend.dto.result.ShowCalendarResult;
 import com.woohaengshi.backend.exception.WoohaengshiException;
 import com.woohaengshi.backend.repository.MemberRepository;
-import com.woohaengshi.backend.repository.StatisticsRepository;
+import com.woohaengshi.backend.repository.statistics.StatisticsRepository;
 import com.woohaengshi.backend.repository.StudySubjectRepository;
 import com.woohaengshi.backend.repository.SubjectRepository;
 import com.woohaengshi.backend.repository.studyrecord.StudyRecordRepository;
