+++ src/test/java/com/woohaengshi/backend/service/StudyRecordServiceTest.java
@@ -22,7 +22,7 @@ import com.woohaengshi.backend.dto.result.ShowCalendarResult;
 import com.woohaengshi.backend.dto.result.SubjectResult;
 import com.woohaengshi.backend.exception.WoohaengshiException;
 import com.woohaengshi.backend.repository.MemberRepository;
-import com.woohaengshi.backend.repository.StatisticsRepository;
+import com.woohaengshi.backend.repository.statistics.StatisticsRepository;
 import com.woohaengshi.backend.repository.StudySubjectRepository;
 import com.woohaengshi.backend.repository.SubjectRepository;
 import com.woohaengshi.backend.repository.studyrecord.StudyRecordRepository;
