+++ src/test/java/com/woohaengshi/backend/service/auth/AuthServiceTest.java
@@ -18,7 +18,7 @@ import com.woohaengshi.backend.dto.request.auth.SignUpRequest;
 import com.woohaengshi.backend.exception.WoohaengshiException;
 import com.woohaengshi.backend.repository.MemberRepository;
 import com.woohaengshi.backend.repository.RefreshTokenRepository;
-import com.woohaengshi.backend.repository.StatisticsRepository;
+import com.woohaengshi.backend.repository.statistics.StatisticsRepository;
 import com.woohaengshi.backend.repository.SubjectRepository;
 import com.woohaengshi.backend.support.fixture.MemberFixture;
